app2/requirements.txt
Django~=4.0.6
djangorestframework~=3.13.1
pytest~=7.1.2
requests~=2.28.1
pika~=1.3.0
redis~=4.3.4
gunicorn